refactor(useNetwork): create `NetworkContext`
