+++ src/hooks/useNetwork/index.ts
@@ -1,3 +1,5 @@
 import useNetwork from './useNetwork';
+import NetworkProvider from './useNetwork.context';
 
 export default useNetwork;
+export { NetworkProvider };
