+++ src/hooks/useNetwork/useNetwork.type.ts
@@ -1,0 +1,16 @@
+import { NetworkConfig } from 'config/environment';
+import { Network } from 'types/network';
+
+import { SetValue } from 'hooks/useLocalStorage';
+
+declare global {
+  interface Window {
+    ethereum: any;
+  }
+}
+
+export type NetworkContextState = {
+  network: Network;
+  networkConfig: NetworkConfig;
+  setNetwork: SetValue<string>;
+};
